Editing

+++ Accounts/DataLayer/AdoNet/PayrollPayElementDao.vb
@@ -70,7 +70,7 @@ Namespace DataLayer.AdoNet
             .EmployeeIdNo = AATM.DataLayer.AdoNet.Extensions.AsId(Of Int32)(reader("EmployeeIdNo")),
             .IdNo = AATM.DataLayer.AdoNet.Extensions.AsId(Of Int32)(reader("IdNo")),
             .PayrollDetailIdNo = AATM.DataLayer.AdoNet.Extensions.AsId(Of Int32)(reader("PayrollDetailIdNo")),
-            .PayrollIdNo = AATM.DataLayer.AdoNet.Extensions.AsId(Of Int32)(reader("PayrollIdNo")),
+            .PayrollIdNo = AATM.DataLayer.AdoNet.Extensions.AsId(Of Int16)(reader("PayrollIdNo")),
             .RecurringPayElementIdNo = AATM.DataLayer.AdoNet.Extensions.AsInt(Of Int32)(reader("RecurringPayElementIdNo"))
            }
